Move EntityList to a common folder
